requirements-dev.txt
ruff
pre-commit
pytest
python-dotenv